+++ internal/wire/wire_suite_test.go
@@ -1,6 +1,8 @@
 package wire
 
 import (
+	"bytes"
+
 	"github.com/lucas-clemente/quic-go/internal/protocol"
 	"github.com/lucas-clemente/quic-go/internal/utils"
 	. "github.com/onsi/ginkgo"
@@ -20,6 +22,12 @@ const (
 	// a QUIC version that uses the IETF frame types
 	versionIETFFrames = protocol.VersionTLS
 )
+
+func encodeVarInt(i uint64) []byte {
+	b := &bytes.Buffer{}
+	utils.WriteVarInt(b, i)
+	return b.Bytes()
+}
 
 var _ = BeforeSuite(func() {
 	Expect(utils.GetByteOrder(versionBigEndian)).To(Equal(utils.BigEndian))
